WP4_Thor-main/service/crawlers/WebCrawler/ClearWeb/requirements.txt
Scrapy
html2text
kafka-python
pymongo
urllib3
url_normalize
